// plugins/github-actions/src/apis/builds/types.ts
/*
 * Copyright 2020 Spotify AB
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum BuildStatus {
  Null,
  Success,
  Failure,
  Pending,
  Running,
}

export type Build = {
  commitId: string;
  message: string;
  branch: string;
  status: BuildStatus;
  uri: string;
};

export type BuildDetails = {
  build: Build;
  author: string;
  logUrl: string;
  overviewUrl: string;
};